requirements.txt
matplotlib
numpy
pynvml
pyyaml
scipy
seaborn
tensorboardx
torchmetrics
transformers
tqdm
scikit-image
yapf==0.40.1